rename and fix minor bug
